Minor refactoring.
