+++ src/main/java/org/eclipse/cargotracker/domain/model/cargo/Cargo.java
@@ -55,9 +55,9 @@ import org.eclipse.cargotracker.domain.shared.DomainObjectUtils;
  * the cargo is (on board carrier, in port etc), are captured in this aggregate.
  */
 @Entity
+// TODO [Jakarta EE 8] Apply repeatable annotations.
 @NamedQueries({ @NamedQuery(name = "Cargo.findAll", query = "Select c from Cargo c"),
-		@NamedQuery(name = "Cargo.findByTrackingId", query = "Select c from Cargo c where c.trackingId = :trackingId"),
-		@NamedQuery(name = "Cargo.getAllTrackingIds", query = "Select c.trackingId from Cargo c") })
+		@NamedQuery(name = "Cargo.findByTrackingId", query = "Select c from Cargo c where c.trackingId = :trackingId") })
 public class Cargo implements Serializable {
 
 	private static final long serialVersionUID = 1L;
