+++ src/main/java/org/eclipse/cargotracker/domain/model/cargo/CargoRepository.java
@@ -11,5 +11,4 @@ public interface CargoRepository {
 	void store(Cargo cargo);
 
 	TrackingId nextTrackingId();
-
 }
