+++ src/main/java/org/eclipse/cargotracker/interfaces/tracking/web/Track.java
@@ -28,7 +28,7 @@ import org.eclipse.cargotracker.domain.model.handling.HandlingEventRepository;
  * <p/>
  * In some very simplistic cases, it is fine to not use even an adapter.
  */
-@Named
+@Named("public.track")
 @ViewScoped
 public class Track implements Serializable {
