+++ src/main/webapp/admin/tracking/track.xhtml
@@ -1,91 +1,101 @@
 <html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:h="http://xmlns.jcp.org/jsf/html"
-      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
-      xmlns:p="http://primefaces.org/ui"
-      xmlns:f="http://java.sun.com/jsf/core">
+	xmlns:h="http://xmlns.jcp.org/jsf/html"
+	xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
+	xmlns:p="http://primefaces.org/ui"
+	xmlns:f="http://java.sun.com/jsf/core">
 <ui:composition template="/WEB-INF/templates/common/admin.xhtml">
-    <ui:define name="title">Track Cargo</ui:define>
-    <ui:define name="content">
+	<ui:define name="title">Track Cargo</ui:define>
+	<ui:define name="content">
 
-        <div class="ui-grid-col-2" id="leftCol"></div>
-        <div class="ui-grid-col-8" id="mainCol">
+		<div class="ui-grid-col-2" id="leftCol"></div>
+		<div class="ui-grid-col-8" id="mainCol">
 
-            <br/>
-            <div class="ui-g ui-g-12">
-                <div class="ui-g-6 ui-g-offset-3">
-                    <h:form id="trackingForm">
-                        <h:panelGrid columns="3" cellpadding="15">
-                            <p>Enter your tracking ID:</p>
-                            <p:autoComplete id="trackingIdInput"
-                                            value="#{cargoAdmin.trackingId}"
-                                            required="true"
-                                            forceSelection="true"
-                                            placeholder="XYZ789"
-                                            title="Valid tracking ID"
-                                            completeMethod="#{listCargo.getTrackingIds}"/>
-                            <h:commandButton value="Track!"
-                                                 action="#{cargoAdmin.onTrackById()}"/>
-                        </h:panelGrid>
-                    </h:form>
-                </div>
-            </div>
+			<br />
+			<div class="ui-g ui-g-12">
+				<div class="ui-g-6 ui-g-offset-3">
+					<h:form id="trackingForm">
+						<h:panelGrid columns="3" cellpadding="15">
+							<p>Enter your tracking ID:</p>
+							<p:autoComplete id="trackingIdInput"
+								value="#{admin.track.trackingId}" required="true"
+								forceSelection="true" placeholder="XYZ789"
+								title="Valid tracking ID"
+								completeMethod="#{admin.track.getTrackingIds}" />
+							<h:commandButton value="Track!"
+								action="#{admin.track.onTrackById()}" />
+						</h:panelGrid>
+					</h:form>
+				</div>
+			</div>
 
-            <div class="ui-g-12"><p:separator/></div>
+			<div class="ui-g-12">
+				<p:separator />
+			</div>
 
-            <div class="ui-g-12">
-                <h:panelGroup rendered="#{not empty cargoAdmin.cargoTrack}">
-                    <div id="result">
+			<div class="ui-g-12">
+				<h:panelGroup rendered="#{not empty admin.track.cargo}">
+					<div id="result">
 
-                        <p>Cargo <strong><h:link value="#{cargoAdmin.trackingId}" id="trackingId"
-                                                 outcome="/admin/show.xhtml?faces-redirect=true"
-                                                 style="text-decoration: none;color: black;">
-                            <f:param name="trackingId" value="#{cargoAdmin.trackingId}" />
-                        </h:link></strong> is currently <strong>#{cargoAdmin.cargoTrack.statusText}</strong></p>
+						<p>
+							Cargo <strong><h:link value="#{admin.track.trackingId}"
+									id="trackingId" outcome="/admin/show.xhtml?faces-redirect=true"
+									style="text-decoration: none;color: black;">
+									<f:param name="trackingId" value="#{admin.track.trackingId}" />
+								</h:link></strong> is currently <strong>#{admin.track.cargo.statusText}</strong>
+						</p>
 
-                        <p:tooltip id="toolTipFade" for="trackingId"
-                                   value="Click to see itinerary's details." />
+						<p:tooltip id="toolTipFade" for="trackingId"
+							value="Click to see itinerary's details." />
 
-                        <p>Estimated time of arrival in
-                            <strong>#{cargoAdmin.cargoTrack.destination}</strong>: #{cargoAdmin.cargoTrack.eta}</p>
-                        <ui:fragment rendered="#{not empty cargoAdmin.cargoTrack.nextExpectedActivity}">
-                            <p>#{cargoAdmin.cargoTrack.nextExpectedActivity}</p></ui:fragment>
+						<p>
+							Estimated time of arrival in <strong>#{admin.track.cargo.destination}</strong>:
+							#{admin.track.cargo.eta}
+						</p>
+						<ui:fragment
+							rendered="#{not empty admin.track.cargo.nextExpectedActivity}">
+							<p>#{admin.track.cargo.nextExpectedActivity}</p>
+						</ui:fragment>
 
-                        <h:panelGroup rendered="#{cargoAdmin.cargoTrack.misdirected}">
-                            <p><i class="fa fa-exclamation-triangle"
-                                  style="color: #9b2332"></i> Cargo is misdirected.</p>
-                        </h:panelGroup>
-                    </div>
-                </h:panelGroup>
-            </div>
+						<h:panelGroup rendered="#{admin.track.cargo.misdirected}">
+							<p>
+								<i class="fa fa-exclamation-triangle" style="color: #9b2332"></i>
+								Cargo is misdirected.
+							</p>
+						</h:panelGroup>
+					</div>
+				</h:panelGroup>
+			</div>
 
-            <div class="ui-g ui-g-12">
-                <div class="ui-g-3">
-                    <p:outputLabel value="Handling History"
-                                   style="font-size: 2em; font-weight: bold"/>
-                </div>
+			<div class="ui-g ui-g-12">
+				<div class="ui-g-3">
+					<p:outputLabel value="Handling History"
+						style="font-size: 2em; font-weight: bold" />
+				</div>
 
-                <div class="ui-g-12">
-                    <p:separator/>
-                    <br/><br/>
-                </div>
+				<div class="ui-g-12">
+					<p:separator />
+					<br />
+					<br />
+				</div>
 
-                <h:outputLabel value="There are no events recorded for the tracking #{cargoAdmin.trackingId}."
-                               rendered="#{empty cargoAdmin.cargoTrack.events}"/>
+				<h:outputLabel
+					value="There are no events recorded for the tracking #{admin.track.trackingId}."
+					rendered="#{empty admin.track.cargo.events}" />
 
-                <div class="ui-g-12">
-                    <ui:repeat value="#{cargoAdmin.cargoTrack.events}" var="leg">
-                        <ui:fragment rendered="#{leg.expected}">
+				<div class="ui-g-12">
+					<ui:repeat value="#{admin.track.cargo.events}" var="leg">
+						<ui:fragment rendered="#{leg.expected}">
                             &#160;&#160;<i class="fa fa-check"
-                                           style="color: green"/>&#160;#{leg.description}<br/>
-                        </ui:fragment>
-                        <ui:fragment rendered="#{!leg.expected}">
+								style="color: green" />&#160;#{leg.description}<br />
+						</ui:fragment>
+						<ui:fragment rendered="#{!leg.expected}">
                             &#160;&#160;<em class="fa fa-flag"
-                                           style="color: #9b2332"/>&#160;#{leg.description}<br/>
-                        </ui:fragment>
-                    </ui:repeat>
-                </div>
-            </div>
-        </div>
-    </ui:define>
+								style="color: #9b2332" />&#160;#{leg.description}<br />
+						</ui:fragment>
+					</ui:repeat>
+				</div>
+			</div>
+		</div>
+	</ui:define>
 </ui:composition>
 </html>
